pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astra-agents"
version = "0.1.0"
description = "AI Agent Framework - Forked from Datawhale Hello-Agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Jason Wu", email = "jasonwu001t@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
]

[project.optional-dependencies]
search = [
    "tavily-python>=0.3.0",
    "google-search-results>=2.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = [
    "astra-agents[search,dev]",
]

[project.urls]
Homepage = "https://github.com/jasonwu001t/astra"
Repository = "https://github.com/jasonwu001t/astra"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
